+++ src/_models/Schedule.ts
@@ -1,8 +1,11 @@
+import { UsageRangeForVisitor } from './Room';
+
 // 一般/フロント 共通
 export type Schedule = {
   roomId: string;
   roomName: string;
   roomEmail: string;
+  usageRange: UsageRangeForVisitor;
   scheduleItems: ScheduleItem[];
   eventsIndex: number[];
 };